refactor: enhance chat functionality and improve navigation; add user avatars and loading state management
