+++ lib/main.dart
@@ -1,6 +1,8 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:paysa/APIs/firestore_apis.dart';
 import 'package:paysa/Controllers/authentication_controller.dart';
 import 'package:paysa/Utils/sizes.dart';
 import 'package:paysa/app.dart';
@@ -12,6 +14,11 @@ void main() async {
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
 
   final authController = Get.put(AuthenticationController());
+
+  if (FirebaseAuth.instance.currentUser != null) {
+    authController.user.value = await PFirestoreAPIs.getUser();
+  }
+
   // defining screen size for mobile
   PSize.screenWidth = 420;
   PSize.screenHeight = 840;
